Move csrf stuff back into one test
